LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-en-expected.txt

FAIL navigator.language should be en. Was en-us.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
